third_party/blink/web_tests/platform/generic/fast/css/line-height-zoom-get-computed-style-expected.txt

This tests that page zoom and getComputedStyle of line-height works well together. This test requires testRunner to run. To test manually, open this page in a browser window, and zoom in & reload for about 10 times. This text should get bigger at each step proportionately and the line-height should be constant.

Line height at zoom level0 : 19.20
Line height at zoom level1 : 19.20
Line height at zoom level2 : 19.20
Line height at zoom level3 : 19.20
Line height at zoom level4 : 19.20
Line height at zoom level5 : 19.20
Line height at zoom level6 : 19.20
Line height at zoom level7 : 19.20
Line height at zoom level8 : 19.20
Line height at zoom level9 : 19.20
